refactor: move `Logger` to `logging` subpackage
