+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/core/API.kt
@@ -2,6 +2,7 @@ package org.firstinspires.ftc.teamcode.core
 
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode
 import com.qualcomm.robotcore.eventloop.opmode.OpMode
+import org.firstinspires.ftc.teamcode.core.logging.Logger
 import org.firstinspires.ftc.teamcode.utils.isNotNull
 
 /**
@@ -95,6 +96,6 @@ abstract class API {
     }
 
     companion object {
-        private val log = Logging.Logger(this)
+        private val log = Logger(this)
     }
 }
